Add gradle plugin stub
The plugin connects to a running qmstr-master and sends some remote log
messages.
The project is build with gradle and compiles the gRPC java code from
the provided proto files (those are copied from the qmstr repository and
we need to find a solution to share those files).
For convenience the build can create idea project files and install the resulting
plugin in your local maven repository.